requirements.txt
Django==1.8.3
django-extensions==1.5.5
ipython==3.2.1
djangorestframework==3.1.3
mysqlclient==1.3.6
django-annoying==0.8.3
gunicorn==19.3.0
whitenoise==2.0.2
django-toolbelt==0.0.1
django-rest-auth==0.4.0
django-allauth==0.22.0
django-webpack-loader==0.1.2.1
